requirements.txt
feedparser>=6.0.11
requests>=2.31.0
transformers>=4.33.0
python-dateutil>=2.8.2
pyyaml>=6.0
torch>=2.2.0
spacy>=3.7.0
